split and added environment.c to utils
