+++ src/execution/execute_cmd_list.c
@@ -6,7 +6,7 @@
 /*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/05 16:21:59 by hrother           #+#    #+#             */
-/*   Updated: 2024/04/30 19:22:21 by aprevrha         ###   ########.fr       */
+/*   Updated: 2024/05/06 21:43:47 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,12 +53,14 @@ int	exec_cmd_list(t_list *cmd_list, t_evars *evars)
 {
 	t_list	*current_cmd;
 
-	if (setup_pipes(cmd_list) != EXIT_SUCCESS || redirs_to_fds(cmd_list, evars) != EXIT_SUCCESS)
+	if (setup_pipes(cmd_list) != EXIT_SUCCESS || redirs_to_fds(cmd_list,
+			evars) != EXIT_SUCCESS)
 		return (ft_lstclear(&cmd_list, free_cmd), EXIT_FAILURE);
 	current_cmd = cmd_list;
 	while (current_cmd != NULL)
 	{
-		exec_cmd((t_cmd *)current_cmd->content, cmd_list, &(evars->envp), evars->status);
+		exec_cmd((t_cmd *)current_cmd->content, cmd_list, &(evars->envp),
+			evars->status);
 		current_cmd = current_cmd->next;
 	}
 	ft_lstiter(cmd_list, close_fds);
