+++ src/utils/environment_1.c
@@ -1,34 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   environment.c                                      :+:      :+:    :+:   */
+/*   environment_1.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: hrother <hrother@student.42vienna.com>     +#+  +:+       +#+        */
+/*   By: aprevrha <aprevrha@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/06 17:13:24 by hrother           #+#    #+#             */
-/*   Updated: 2024/04/30 18:03:22 by hrother          ###   ########.fr       */
+/*   Updated: 2024/05/06 22:08:47 by aprevrha         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
-
-void	print_env(void *content)
-{
-	t_env	*env;
-
-	env = (t_env *)content;
-	log_msg(DEBUG, "%s=%s", env->key, env->value);
-}
-
-void	free_env(void *content)
-{
-	t_env	*env;
-
-	env = (t_env *)content;
-	free(env->key);
-	free(env->value);
-	free(env);
-}
+#include "../../includes/minishell.h"
 
 char	*combine_key_value(char *key, char *value)
 {
@@ -41,18 +23,6 @@ char	*combine_key_value(char *key, char *value)
 	result = ft_strjoin(tmp, value);
 	free(tmp);
 	return (result);
-}
-
-t_env	*new_env(char *key, char *value)
-{
-	t_env	*env;
-
-	env = (t_env *)malloc(sizeof(t_env));
-	if (env == NULL)
-		return (NULL);
-	env->key = key;
-	env->value = value;
-	return (env);
 }
 
 char	*get_value(const char *arg)
@@ -75,27 +45,6 @@ char	*get_key(const char *arg)
 	while (arg[i] && arg[i] != '=')
 		i++;
 	return (ft_substr(arg, 0, i));
-}
-
-char	*ft_getenv(t_list *envlst, char *key)
-{
-	t_env	*env;
-	int		len_key;
-	int		len_env_key;
-
-	if (key == NULL)
-		return (NULL);
-	len_key = ft_strlen(key);
-	log_msg(DEBUG, "search key: %s", key);
-	while (envlst != NULL)
-	{
-		env = (t_env *)envlst->content;
-		len_env_key = ft_strlen(env->key);
-		if (len_key == len_env_key && ft_strncmp(env->key, key, len_key) == 0)
-			return (env->value);
-		envlst = envlst->next;
-	}
-	return (NULL);
 }
 
 t_list	*envp_to_list(char **envp)
